Pick up size_t def from Windows rather than making one up and possibly clashing with other packages.
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk@822963 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/include/qpid/sys/windows/IntegerTypes.h
@@ -21,6 +21,8 @@
  *
  */
 
+#include <BaseTsd.h>    /* Windows system types */
+
 typedef unsigned char    uint8_t;
 typedef char             int8_t;
 typedef unsigned short   uint16_t;
@@ -32,7 +34,7 @@ typedef          __int64 int64_t;
 
 // Visual Studio doesn't define other common types, so set them up here too.
 typedef int              pid_t;
-typedef int              ssize_t;
+typedef SSIZE_T          ssize_t;
 typedef unsigned int     uint;
 
 #endif  /*!QPID_SYS_WINDOWS_INTEGERTYPES_H*/
